task 1

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,3 +1,5 @@
+#include <stdlib.h>
+#include "function_pointers.h"
 /**
  * array_iterator - a function given as a
  * parameter on each element of an array.
@@ -10,7 +12,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned int i;
+	size_t i;
 
 	if (!array || !action)
 		return;
